Loop logic

+++ SavvyDB/Models/Order.cs
@@ -6,7 +6,7 @@ namespace SavvyDB.Models
     {
         public int OrderId {get; set;}
         public int CustomerId {get; set;}
-        public int Totalprice { get; set; }
+        public decimal Totalprice { get; set; }
         public DateTime Date { get; set; }
     }
 }
